+++ Magic/CHANGELOG.md
@@ -7,6 +7,7 @@
  - Add light and time requirements, can be used in CheckRequirements or Selector for time of day or light level checks
  - Make it easier to change the items used for spell and wand icons (see items/wands.yml)
  - Fix cycle_reverse wand action causing errors
+ - Cycle mode wands show spell lore by default (use show_cycle_lore: false in wand properties to turn this off)
 
 # 7.7.5